複数画像の投稿・削除完了

+++ app/Http/Controllers/AnswerController.php
@@ -29,24 +29,25 @@ class AnswerController extends Controller
     public function create(Request $request)
 	{
 		$this->validate($request, Answer::$rules);
-
+	
 		$answer = new Answer;
 		$form = $request->all();
 		$answer->user_id = Auth::user()->id;
-		$answer->question_id = $request->question_id;
-	
+
 		unset($form['_token']);
 		unset($form['image_paths']);
 	
 		$answer->fill($form);
 		$answer->save();
 
-		foreach ($request->file('image_paths') as $index => $e) {
-			$picture = new Picture;
-			$path = Storage::disk('s3')->putFile('/',$e,'public');
-			$picture->image_path = Storage::disk('s3')->url($path);
-			$picture->answer_id = $answer->id;
-			$picture->save();
+		if(!empty($request->file('image_paths'))){
+			foreach ($request->file('image_paths') as $index => $e) {
+				$picture = new Picture;
+				$path = Storage::disk('s3')->putFile('/',$e,'public');
+				$picture->image_path = Storage::disk('s3')->url($path);
+				$picture->answer_id = $answer->id;
+				$picture->save();
+			}
 		}
 	
 		return redirect('/answer/create/'.$answer->question_id)->with('message', '回答ありがとうございます。');
@@ -65,8 +66,8 @@ class AnswerController extends Controller
 		if($answer->user_id != Auth::user()->id){
 			abort(404);
 		}
-		$question = Question::find($answer->question_id);
-		$user = User::find($question->user_id);
+		$question = Question::findOrFail($answer->question_id);
+		$user = User::findOrFail($question->user_id);
 		return view('answer.edit', ['answer' => $answer , 'question' => $question , 'user' => $user]);
 	}
 
@@ -77,26 +78,30 @@ class AnswerController extends Controller
 		$answer = Answer::find($request->id);
 		$answer_form = $request->all();
 
-		foreach ($request->remove as $key => $value) {
-			$image = Picture::find($value);
-			$image_path = explode("/", $image->image_path);
-			$img = end($image_path);
-			$disk = Storage::disk('s3');
-			$disk->delete($img);
-			$image->delete();
-		}
-
 		unset($answer_form['_token']);
+		unset($answer_form['image_paths']);
 		unset($answer_form['remove']);
 		$answer->fill($answer_form)->save();
 
-		foreach ($request->file('image_paths') as $index => $e) {
-			$picture = new Picture;
-			$path = Storage::disk('s3')->putFile('/',$e,'public');
-			$picture->image_path = Storage::disk('s3')->url($path);
-			$picture->answer_id = $answer->id;
-			$picture->save();
-			unset($answer_form['image_paths']);
+		if(!empty($request->file('image_paths'))){
+			foreach ($request->file('image_paths') as $index => $e) {
+				$picture = new Picture;
+				$path = Storage::disk('s3')->putFile('/',$e,'public');
+				$picture->image_path = Storage::disk('s3')->url($path);
+				$picture->answer_id = $answer->id;
+				$picture->save();
+			}
+		}
+
+		if(!empty($request->remove)){
+			foreach ($request->remove as $key => $value) {
+				$image = Picture::find($value);
+				$image_path = explode("/", $image->image_path);
+				$img = end($image_path);
+				$disk = Storage::disk('s3');
+				$disk->delete($img);
+				$image->delete();
+			}
 		}
 
 		return redirect('/list/answers')->with('answeredit', '投稿を変更しました。');
